Yael.Eff.Builder

Documentation

runEffTWith :: EffStack f => Builder f m -> EffT (EffList (ToEffList f)) m a -> m a

Run an effectful computation using a set of effect handlers that are implemented in terms of each other.

wrapping :: (f m -> g m -> g m) -> (f m -> g m) -> f m -> g m infixr 9

Helper datatypes and classes in case you want to write EffStack instances by hand (but hint: you shouldn't)

data EffList fs m where

Constructors

EffNil :: EffList '[] m

EffCons :: f m -> EffList fs m -> EffList (f ': fs) m

class EffStack f where

Minimal complete definition

Nothing

Associated Types

type ToEffList f :: [(* -> *) -> *]

Methods

build :: Builder f m -> EffList (ToEffList f) m

build :: (GTo (Rep (f (EffectIdentity m))) (Rep (f EffectProxy)) m, GApply (Rep (Builder f m)) (Rep (f (EffectIdentity m))) (GToEffList (Rep (f EffectProxy))) m, Generic (f (EffectIdentity m)), Generic (Builder f m), ToEffList f ~ GToEffList (Rep (f EffectProxy))) => Builder f m -> EffList (ToEffList f) m

toEffList :: f (EffectIdentity m) -> EffList (ToEffList f) m

toEffList :: (Generic (f (EffectIdentity m)), GTo (Rep (f (EffectIdentity m))) (Rep (f EffectProxy)) m, ToEffList f ~ GToEffList (Rep (f EffectProxy))) => f (EffectIdentity m) -> EffList (ToEffList f) m

data EffectIdentity m

data EffectBuilder s (m :: * -> *)

data EffectProxy

type family Effect f t where ...

Equations

Effect (EffectIdentity m) t = t m

Effect (EffectBuilder s m) t = EffList (ToEffList s) m -> t m

Effect EffectProxy t = Proxy t